refactor: Move SecretOnApplicationStart method to Application.xaml.cs for better organization
